+++ jax/_src/sharding_impls.py
@@ -217,7 +217,8 @@ class NamedSharding(XLACompatibleSharding):
     # representation of Parsed Pspec
     if self._parsed_pspec is None:
       self._parsed_pspec, _, _ = prepare_axis_resources(
-          self.spec, "NamedSharding spec", allow_unconstrained_dims=True)
+          PartitionSpec() if self.spec is None else self.spec,
+          "NamedSharding spec", allow_unconstrained_dims=True)
 
     _check_mesh_resource_axis(self.mesh, self._parsed_pspec)
 
@@ -956,7 +957,7 @@ def prepare_axis_resources(axis_resources,
 
   new_entries = []
   for entry in entries:
-    if is_unspecified_or_auto(entry):
+    if is_unspecified_or_auto(entry) or entry is None:
       new_entries.append(entry)
     elif isinstance(entry, sharding.Sharding):
       if isinstance(entry, PmapSharding):